+++ src/app/app.module.ts
@@ -6,8 +6,8 @@ import { OAuthModule } from 'angular-oauth2-oidc';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { AuthorizationInterceptor } from './interceptors/authorization.interceptor';
-import { DashboardComponent } from './dashboard/dashboard.component';
-import { AnonymousComponent } from './anonymous/anonymous.component';
+import { DashboardComponent } from './components/dashboard/dashboard.component';
+import { AnonymousComponent } from './components/anonymous/anonymous.component';
 
 @NgModule({
   declarations: [
