+++ src/app/components/anonymous/anonymous.component.ts
